So it seems the watch options for collect-comment-tests is broken now, as it actually imports modules. So we need to run a new process each time.
