+++ src/test/reporting.ts
@@ -7,6 +7,7 @@ export type TestReport = {
   name: string;
   failed: boolean;
   dot: { snapshot: string; current: string };
+  svg: { snapshot: string; current: string };
   code: string;
   results: TestResults[];
 };
